Introduced a UDP connection to our azure iot hub
a new process data subvi has been input into the read data message handling loop. It will read our data cluster in with the host name, remote and local ports as well as the key and connection id. Within this subvi we also call the encryptor dll to encrypt our data being sent to the azure hub. The json payload is passed in from our string that we have styled into json format